feat: use ld-audit-search-mod instead of LD_FALLBACK
Replaces the use of LD_FALLBACK with a new tool called
ld-audit-search-mod. Instead of requiring a patch to a _very_ low level
component of the GNU software stack, it instead relies on the built-in
LD_AUDIT mechanism to redirect dynamic loads at runtime. It is also more
configurable than the old patch, but I've just set it up to emulate the
fallback behavior for now.